+++ lib/src/simple_dynamic_form.dart
@@ -1,16 +1,15 @@
-import 'package:dynamic_form/dynamic_form.dart';
-import 'package:dynamic_form/src/card_number_field.dart';
-import 'package:dynamic_form/src/country_text_field.dart';
-import 'package:dynamic_form/src/cvv_text_field.dart';
-import 'package:dynamic_form/src/date_input_field.dart';
-import 'package:dynamic_form/src/date_text_field.dart';
-import 'package:dynamic_form/src/element.dart';
-import 'package:dynamic_form/src/email_text_field.dart';
-import 'package:dynamic_form/src/group_elements.dart';
-import 'package:dynamic_form/src/password_text_field.dart';
-import 'package:dynamic_form/src/phone_text_field.dart';
-import 'package:dynamic_form/src/text_area_form_field.dart';
-import 'package:dynamic_form/src/utilities/constants.dart';
+import './widgets/card_number_field.dart';
+import './widgets/country_text_field.dart';
+import './widgets/cvv_text_field.dart';
+import './widgets/date_input_field.dart';
+import './widgets/date_text_field.dart';
+import './elements/element.dart';
+import './widgets/email_text_field.dart';
+import './elements/group_elements.dart';
+import './widgets/password_text_field.dart';
+import './widgets/phone_text_field.dart';
+import './widgets/text_area_form_field.dart';
+import './utilities/constants.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
