+++ lib/src/widgets/date_input_field.dart
@@ -2,8 +2,8 @@ import 'package:dynamic_form/dynamic_form.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
-import 'utilities/constants.dart';
-import 'utilities/text_controller_format_input.dart';
+import '../utilities/constants.dart';
+import '../utilities/text_controller_format_input.dart';
 
 class DateInputField extends StatelessWidget {
   final DateInputElement element;
