+++ pubspec.yaml
@@ -11,7 +11,7 @@ dependencies:
     sdk: flutter
   flag: ^3.1.0
   http: ^0.12.2
-  flutter_svg: ^0.19.0
+  flutter_svg: ^0.19.1
   intl: ^0.16.1
 dev_dependencies:
   flutter_test:
